Extrapolate dialogs to load csv-based oq-engine outputs
