+++ svir/dialogs/drive_oq_engine_server_dialog.py
@@ -62,6 +62,10 @@ from svir.utilities.utils import (WaitCursorManager,
                                   SvNetworkError,
                                   )
 from svir.dialogs.load_output_as_layer_dialog import LoadOutputAsLayerDialog
+from svir.dialogs.load_ruptures_as_layer_dialog import (
+    LoadRupturesAsLayerDialog)
+from svir.dialogs.load_dmg_by_asset_as_layer_dialog import (
+    LoadDmgByAssetAsLayerDialog)
 from svir.dialogs.show_full_report_dialog import ShowFullReportDialog
 
 FORM_CLASS = get_ui_class('ui_drive_engine_server.ui')
@@ -481,8 +485,15 @@ class DriveOqEngineServerDialog(QDialog, FORM_CLASS):
             if outtype in ('npz', 'csv'):
                 filepath = self.download_output(
                     output_id, outtype, dest_folder)
-                dlg = LoadOutputAsLayerDialog(
-                    self.iface, output_type, filepath)
+                if output_type == 'ruptures':
+                    dlg = LoadRupturesAsLayerDialog(
+                        self.iface, output_type, filepath)
+                elif output_type == 'dmg_by_asset':
+                    dlg = LoadDmgByAssetAsLayerDialog(
+                        self.iface, output_type, filepath)
+                else:
+                    dlg = LoadOutputAsLayerDialog(
+                        self.iface, output_type, filepath)
                 dlg.exec_()
             else:
                 raise NotImplementedError("%s %s" % (action, outtype))
